plugins/PY/requirements.txt
bottle==0.12.25
fastapi==0.104.1
flask==3.0.0
httpx==0.25.1
mysql-connector-python==8.2.0
pymongo==4.6.0
PyMySQL==1.1.0
redis==5.0.1
requests==2.31.0
SQLAlchemy==2.0.23
starlette-context==0.3.6
mysqlclient==2.2.0